Set weigth as decimal
As in DB.

+++ Models/AutofillerModel.cs
@@ -82,7 +82,7 @@ namespace Autofiller.Models
         public string value { get; set; }
         public DateTime created_at { get; set; }
         public DateTime updated_at { get; set; }
-        public double weight { get; set; }
+        public decimal weight { get; set; }
 
         //Foreign Keys
         public virtual Pivot Pivot { get; set; }
